# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: gotomemory Gateway API
  version: 0.1.0
  description: >
    Cross-platform Memory Control Plane. This document is the single source of truth
    for the API (system spec §9) and data model (§8). TS and Python clients are
    generated from it — see monorepo-guide §6. Privacy constraints are enforced at the
    schema level: search responses never include `content`, and decision objects never
    include memory bodies, full summaries, or raw embeddings.
servers:
  - url: https://api.gotomemory.example/v1
    description: Gateway API v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Create a user session from a provider credential.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginRequest" }
      responses:
        "201":
          description: Authenticated session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthLoginResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Read the current authenticated user.
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthMeResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke the current session token.
      responses:
        "204": { description: Session revoked. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /memories:
    post:
      tags: [memories]
      operationId: createMemory
      summary: Create a memory (§9.1)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMemoryRequest" }
      responses:
        "201":
          description: Created (or pending confirmation). Returns classifier/policy-resolved fields.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateMemoryResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "422": { $ref: "#/components/responses/ClassificationRequired" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /memories/search:
    post:
      tags: [memories]
      operationId: searchMemories
      summary: Search memories — previews only, never `content` (§9.2)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: Ranked preview results with per-item access decisions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /memories/{id}:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    get:
      tags: [memories]
      operationId: readMemory
      summary: Read full memory content (§9.2.1) — requires memory:read + purpose
      parameters:
        - name: purpose
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full content plus current version (ETag carries the version too).
          headers:
            ETag:
              schema: { type: string }
              description: Opaque version tag for optimistic concurrency.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryRead" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [memories]
      operationId: updateMemory
      summary: Update a memory with optimistic locking (§9.4)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMemoryRequest" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateMemoryResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
    delete:
      tags: [memories]
      operationId: deleteMemory
      summary: Soft-delete; de-index immediately, purge later (§9.5)
      responses:
        "204": { description: Accepted — de-indexed immediately; physical purge deferred. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /context/build:
    post:
      tags: [context]
      operationId: buildContext
      summary: Build model context — the governance API (§9.3)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContextBuildRequest" }
      responses:
        "200":
          description: Injectable context plus decision_id, memory_ids, and omitted reasons.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextBuildResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /context/confirm:
    post:
      tags: [context]
      operationId: confirmContext
      summary: Redeem a one-time confirmation token for injectable context (§9.3.1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContextConfirmRequest" }
      responses:
        "200":
          description: Final injectable context after confirmation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextBuildResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /pages:
    post:
      tags: [pages]
      operationId: createPage
      summary: Publish a read-only shared page.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePageRequest" }
      responses:
        "201":
          description: Published page metadata and share URL.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
    get:
      tags: [pages]
      operationId: listPages
      summary: List shared pages owned by the current user.
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Page list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageListResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /pages/public/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [pages]
      operationId: getPublicPage
      summary: Read public shared page data for frontend rendering.
      responses:
        "200":
          description: Shared page metadata and artifact content.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicPageResponse" }
        "403": { $ref: "#/components/responses/PolicyDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /pages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [pages]
      operationId: getPage
      summary: Read shared page metadata.
      responses:
        "200":
          description: Page metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [pages]
      operationId: updatePage
      summary: Update shared page metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePageRequest" }
      responses:
        "200":
          description: Updated page metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
    delete:
      tags: [pages]
      operationId: unpublishPage
      summary: Unpublish a shared page.
      responses:
        "204": { description: Page unpublished. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /pages/{id}/versions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [pages]
      operationId: createPageVersion
      summary: Publish a new read-only version for an existing shared page.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePageVersionRequest" }
      responses:
        "201":
          description: Updated page metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: OAuth 2.0 access token (users/CLI) or tenant API token (services). See §13.2.
  parameters:
    MemoryId:
      name: id
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Required for auto-write and bulk import to dedupe repeated submissions.
  responses:
    InvalidRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthenticated:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PolicyDenied:
      description: Denied by Memory Policy; body never includes protected content.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    VersionConflict:
      description: Optimistic-lock version mismatch.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ClassificationRequired:
      description: Candidate could not be classified / needs confirmation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Sensitivity:
      type: string
      description: Ordered public < normal < private < secret (§8.3).
      enum: [public, normal, private, secret]
    EmbeddingPolicy:
      type: string
      enum: [allowed, redacted_only, disabled]
    Freshness:
      type: string
      enum: [current_state, historical_fact, timeless, temporary]
    MemoryStatus:
      type: string
      enum: [active, superseded, expired, deleted, pending_confirmation]
    MemoryType:
      type: string
      enum: [preference, fact, instruction, relationship, workflow, credential_hint, note]
    Scope:
      type: string
      enum: [personal, project, team, session]
    Source:
      type: string
      enum: [user_explicit, manual, chatgpt, claude, gemini, import, api]
    Platform:
      type: string
      enum: [chatgpt, claude, gemini]
...
    PageKind:
      type: string
      enum: [html, markdown, pdf, docx, xlsx, pptx]
    PageVisibility:
      type: string
      enum: [private, unlisted, public]
    PageStatus:
      type: string
      enum: [active, unpublished, expired, deleted, quarantined]
    PageSource:
      type: string
      enum: [api, mcp, cli, console, agent, import]
    ExpiryUnit:
      type: string
      enum: [hours, days]
    AuthProvider:
      type: string
      enum: [google, github]

    # --- requests / responses ---
    AuthLoginRequest:
      type: object
      required: [provider, provider_user_id, email, name, mock_access_token]
      properties:
        provider: { $ref: "#/components/schemas/AuthProvider" }
        provider_user_id: { type: string }
        email: { type: string }
        name: { type: string }
        avatar_url: { type: string }
        mock_access_token:
          type: string
          description: Temporary provider credential used by the local mock OAuth buttons.
      additionalProperties: false
    AuthUser:
      type: object
      required: [id, tenant_id, provider, provider_user_id, email, name]
      properties:
        id: { type: string }
        tenant_id: { type: string }
        provider: { $ref: "#/components/schemas/AuthProvider" }
        provider_user_id: { type: string }
        email: { type: string }
        name: { type: string }
        avatar_url: { type: string }
      additionalProperties: false
    AuthLoginResponse:
      type: object
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/AuthUser" }
      additionalProperties: false
    AuthMeResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
      additionalProperties: false
    CreateMemoryRequest:
      type: object
      required: [scope, type, content, source]
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        type: { $ref: "#/components/schemas/MemoryType" }
        content: { type: string }
        summary: { type: string }
        subject: { type: string }
        predicate: { type: string }
        value: { type: string }
        tags:
          type: array
          items: { type: string }
        source: { $ref: "#/components/schemas/Source" }
        sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        collection_id: { type: string }
        ttl: { type: string, format: date-time }
      additionalProperties: false
    CreateMemoryResponse:
      type: object
      required: [id, status, version]
      description: >
        Returns classifier/policy-resolved fields, since submitted sensitivity may be
        upgraded and status may land on pending_confirmation (§9.1).
      properties:
        id: { type: string }
        status: { $ref: "#/components/schemas/MemoryStatus" }
        sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        summary_sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        embedding_policy: { $ref: "#/components/schemas/EmbeddingPolicy" }
        version: { type: integer }
        confirmation: { $ref: "#/components/schemas/Confirmation" }
      additionalProperties: false
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        platform: { $ref: "#/components/schemas/Platform" }
        client_id: { type: string }
        purpose: { type: string }
        scope:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
        project_id: { type: string }
        limit: { type: integer, description: "Max results (server default 12)." }
        cursor:
          type: [string, "null"]
      additionalProperties: false
    SearchResultItem:
      type: object
      description: Preview-only result. Never includes `content` (§9.2, §17.5).
      required: [id, summary_preview, sensitivity, version, score, access]
      properties:
        id: { type: string }
        summary_preview: { type: string }
        sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        version: { type: integer }
        score: { type: number }
        access: { $ref: "#/components/schemas/AccessDecision" }
      additionalProperties: false
    AccessDecision:
      type: object
      required: [can_read_content, can_inject, requires_confirmation]
      properties:
        can_read_content: { type: boolean }
        can_inject: { type: boolean }
        requires_confirmation: { type: boolean }
      additionalProperties: false
    SearchResponse:
      type: object
      required: [items, next_cursor, decision_id]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SearchResultItem" }
        next_cursor:
          type: [string, "null"]
        decision_id: { type: string }
      additionalProperties: false
    MemoryRead:
      type: object
      required: [id, content, sensitivity, freshness, status, version, decision_id]
      properties:
        id: { type: string }
        content: { type: string }
        sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        status: { $ref: "#/components/schemas/MemoryStatus" }
        version: { type: integer }
        decision_id: { type: string }
      additionalProperties: false
    UpdateMemoryRequest:
      type: object
      required: [version]
      properties:
        content: { type: string }
        summary: { type: string }
        tags:
          type: array
          items: { type: string }
        sensitivity: { $ref: "#/components/schemas/Sensitivity" }
        status: { $ref: "#/components/schemas/MemoryStatus" }
        version: { type: integer, description: Expected current version (optimistic lock). }
      additionalProperties: false
    ContextBuildRequest:
      type: object
      required: [platform, client_id, task]
      properties:
        platform: { $ref: "#/components/schemas/Platform" }
        client_id: { type: string }
        purpose: { type: string }
        task: { type: string }
        conversation_excerpt:
          type: string
          description: Used only for this retrieval; never logged or persisted (§9.3, §18).
        token_budget:
          {
            type: integer,
            description: "Token budget for the built context (server default 1200).",
          }
        project_id: { type: string }
      additionalProperties: false
    ContextBuildResponse:
      type: object
      required: [memory_ids, redacted, requires_confirmation, decision_id, omitted]
      properties:
        context:
          type: [string, "null"]
          description: Injectable context; null when everything requires confirmation.
        memory_ids:
          type: array
          items: { type: string }
        redacted: { type: boolean }
        requires_confirmation: { type: boolean }
        decision_id: { type: string }
        confirmation: { $ref: "#/components/schemas/Confirmation" }
        omitted:
          type: array
          items: { $ref: "#/components/schemas/OmittedMemory" }
      additionalProperties: false
    ContextConfirmRequest:
      type: object
      required: [decision_id, confirmation_token, confirmed_memory_ids]
      properties:
        decision_id: { type: string }
        confirmation_token: { type: string }
        confirmed_memory_ids:
          type: array
          items: { type: string }
      additionalProperties: false
    ExpiresIn:
      type: object
      required: [value, unit]
      properties:
        value: { type: integer, minimum: 1 }
        unit: { $ref: "#/components/schemas/ExpiryUnit" }
      additionalProperties: false
    CreatePageRequest:
      type: object
      required: [title, kind]
      properties:
        title: { type: string }
        kind: { $ref: "#/components/schemas/PageKind" }
        content: { type: string }
        content_base64: { type: string }
        filename: { type: string }
        description: { type: string }
        visibility: { $ref: "#/components/schemas/PageVisibility" }
        expires_in: { $ref: "#/components/schemas/ExpiresIn" }
        ttl_hours:
          type: integer
          description: Deprecated compatibility alias for expires_in hours.
        source: { $ref: "#/components/schemas/PageSource" }
      additionalProperties: false
    UpdatePageRequest:
      type: object
      required: [version]
      properties:
        title: { type: string }
        description: { type: [string, "null"] }
        visibility: { $ref: "#/components/schemas/PageVisibility" }
        expires_at: { type: [string, "null"], format: date-time }
        status: { $ref: "#/components/schemas/PageStatus" }
        version: { type: integer }
      additionalProperties: false
    CreatePageVersionRequest:
      type: object
      required: [version]
      properties:
        content: { type: string }
        content_base64: { type: string }
        filename: { type: string }
        version: { type: integer }
      additionalProperties: false
    PageResponse:
      type: object
      required:
        [
          id,
          slug,
          title,
          description,
          kind,
          url,
          visibility,
          status,
          expires_at,
          created_at,
          updated_at,
          version,
        ]
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        kind: { $ref: "#/components/schemas/PageKind" }
        url: { type: string }
        visibility: { $ref: "#/components/schemas/PageVisibility" }
        status: { $ref: "#/components/schemas/PageStatus" }
        expires_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
      additionalProperties: false
    PageListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PageResponse" }
        next_cursor: { type: [string, "null"] }
      additionalProperties: false
    PublicPageResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          required: [content, filename, mime_type, size_bytes]
          properties:
            content: { type: string }
            filename: { type: [string, "null"] }
            mime_type: { type: string }
            size_bytes: { type: integer }
          additionalProperties: false
    OmittedMemory:
      type: object
      required: [memory_id, reason]
      properties:
        memory_id: { type: string }
        reason:
          type: string
          examples: [policy_denied, sensitivity_exceeds_policy, requires_confirmation]
      additionalProperties: false
    Confirmation:
      type: object
      description: One-time confirmation handle with a redacted preview (§8.7, §9.3.1).
      required: [confirmation_token, expires_at]
      properties:
        confirmation_token: { type: string }
        expires_at: { type: string, format: date-time }
        preview:
          type: array
          items: { $ref: "#/components/schemas/SearchResultItem" }
      additionalProperties: false
...
    Error:
      type: object
      description: Unified error envelope (§9.8). Never contains protected content.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthenticated
                - policy_denied
                - not_found
                - version_conflict
                - classification_required
                - rate_limited
                - unsupported_artifact_type
                - artifact_too_large
                - invalid_artifact
                - share_policy_denied
                - page_not_found
                - page_version_conflict
                - render_failed
                - page_quarantined
                - internal
            message: { type: string }
            decision_id:
              type: [string, "null"]
            details:
              type: object
              additionalProperties: true
          additionalProperties: false
      additionalProperties: false